+++ ft_printf/ft_getprintint.c
@@ -6,24 +6,24 @@
 /*   By: bpoyet <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 13:53:59 by bpoyet            #+#    #+#             */
-/*   Updated: 2023/11/23 16:41:51 by bpoyet           ###   ########.fr       */
+/*   Updated: 2023/11/24 11:17:07 by bpoyet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft/libft.h"
 
-int	ft_getprintint(va_list parameter)
+ssize_t	ft_getprintint(va_list parameter)
 {
 	int		value;
-	int error;
-	int i;
+	int		error;
+	ssize_t	i;
 
 	i = 0;
 	error = 1;
 	value = va_arg(parameter, int);
 	ft_putnbr(value, 1, &error, &i);
-	if(!error)
-		return(-1);
+	if (!error)
+		return (-1);
 	return (i);
 }
