+++ ft_printf/ft_putnbr.c
@@ -6,7 +6,7 @@
 /*   By: bpoyet <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 13:06:57 by bpoyet            #+#    #+#             */
-/*   Updated: 2023/11/23 16:42:04 by bpoyet           ###   ########.fr       */
+/*   Updated: 2023/11/24 11:17:36 by bpoyet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ static int	writenbr(char c, int fd)
 	return (write(fd, &c, 1));
 }
 
-void	ft_putnbr(int n, int fd, int *error, int *i)
+void	ft_putnbr(int n, int fd, int *error, ssize_t *i)
 {
 	unsigned int	absolut;
 
@@ -32,6 +32,6 @@ void	ft_putnbr(int n, int fd, int *error, int *i)
 	if (absolut > 9)
 		ft_putnbr(absolut / 10, fd, error, i);
 	*i = *i + 1;
-	if(*error && writenbr(absolut % 10 + '0', fd) < 0)
+	if (*error && writenbr(absolut % 10 + '0', fd) < 0)
 		*error = 0;
 }
